contoso_agents/modem_check_agent/pyproject.toml
[project]
name = "rogers-modem-check-agent"
version = "1.0.0"
description = "Rogers Modem Check Agent for A2A multi-agent system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.2.6",
    "azure-ai-projects>=1.0.0b12",
    "azure-ai-agents>=1.1.0b2",
    "azure-identity>=1.23.0",
    "uvicorn>=0.34.2",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.35.0",
    "httpx>=0.25.0",
]
